+++ src/to_index.cpp
@@ -147,22 +147,21 @@ void IndexInputVector::initialize(const SEXP &x){
     if(TYPEOF(x) == CHARSXP || TYPEOF(x) == LGLSXP || TYPEOF(x) == INTSXP || 
       TYPEOF(x) == REALSXP || TYPEOF(x) == CPLXSXP || TYPEOF(x) == STRSXP || TYPEOF(x) == RAWSXP){
       // we convert to character
-      SEXP call_as_character = PROTECT(Rf_lang2(Rf_install("as.character"), x));
   
       int any_error;
-      this->x_conv = PROTECT(R_tryEval(call_as_character, R_GlobalEnv, &any_error));
+      this->x_conv = PROTECT(R_tryEval(Rf_lang2(Rf_install("as.character"), x), R_GlobalEnv, &any_error));
+      this->is_protect = true;
 
       if(any_error){
-        Rf_error("In `to_index`, the vector to index was not standard (int or real, etc) and failed to be converted to character before applying indexation._n");
+        Rcpp::stop("In `to_index`, the vector to index was not standard (int or real, etc) and failed to be converted to character before applying indexation._n");
       }
       
       // conversion succeeded
       this->type = T_STR;
       this->px_intptr = (intptr_t *) STRING_PTR_RO(this->x_conv);
-      this->is_protect = true;
       
     } else {
-      Rf_error("In `to_index`, the R vectors must be atomic. The current type is not valid.");
+      Rcpp::stop("In `to_index`, the R vectors must be atomic. The current type is not valid.");
     }    
     
   }
@@ -673,7 +672,7 @@ inline void update_index_intarray_g_obs(int id, size_t i, int &g, int * &int_arr
   }  
 }
 
-void multiple_ints_to_index(const vector<IndexInputVector> &all_vecs, vector<int> &all_k, 
+void multiple_ints_to_index(const vector<std::shared_ptr<IndexInputVector>> &all_pvecs, vector<int> &all_k, 
                             int *__restrict p_index, int &n_groups, const bool is_final,
                             vector<int> &vec_firstobs, vector<int> &vec_table,
                             vector<double> &vec_sum,
@@ -683,11 +682,11 @@ void multiple_ints_to_index(const vector<IndexInputVector> &all_vecs, vector<int
   int K = all_k.size();
     
   for(auto &&k : all_k){
-    sum_bin_ranges += all_vecs[k].x_range_bin;
+    sum_bin_ranges += all_pvecs[k]->x_range_bin;
   }  
   
   int k0 = all_k[0];
-  const IndexInputVector *x0 = &all_vecs[k0];
+  const IndexInputVector *x0 = all_pvecs[k0].get();
   const size_t n = x0->n;
   const int * px0_int = (int *) x0->px_int;
   const double * px0_dbl = (double *) x0->px_dbl;
@@ -751,7 +750,7 @@ void multiple_ints_to_index(const vector<IndexInputVector> &all_vecs, vector<int
   } else {
     
     int k1 = all_k[1];
-    const IndexInputVector *x1 = &all_vecs[k1];
+    const IndexInputVector *x1 = all_pvecs[k1].get();
     const int *px1_int = (int *) x1->px_int;
     const double *px1_dbl = (double *) x1->px_dbl;
     
@@ -842,7 +841,7 @@ void multiple_ints_to_index(const vector<IndexInputVector> &all_vecs, vector<int
       offset += x1->x_range_bin;
       for(int ind=2 ; ind<K-1 ; ++ind){
         int k = all_k[ind];
-        const IndexInputVector *xk = &all_vecs[k];
+        const IndexInputVector *xk = all_pvecs[k].get();
         const int *pxk_int = (int *) xk->px_int;
         const double *pxk_dbl = (double *) xk->px_dbl;
         
@@ -862,7 +861,7 @@ void multiple_ints_to_index(const vector<IndexInputVector> &all_vecs, vector<int
             
       // last element + group creation
       int k = all_k[K - 1];
-      const IndexInputVector *xk = &all_vecs[k];
+      const IndexInputVector *xk = all_pvecs[k].get();
       const int *pxk_int = (int *) xk->px_int;
       const double *pxk_dbl = (double *) xk->px_dbl;
       
@@ -914,40 +913,40 @@ void multiple_ints_to_index(const vector<IndexInputVector> &all_vecs, vector<int
   delete[] int_array;
 }
 
-std::vector<IndexInputVector> SEXP_to_vec_r_vector(const SEXP &x){
+std::vector<std::shared_ptr<IndexInputVector>> SEXP_to_vec_r_vector(const SEXP &x){
   
   size_t n = 0;
   int K = 0;
-  std::vector<IndexInputVector> all_vecs;
+  std::vector<std::shared_ptr<IndexInputVector>> all_pvecs;
   
   // we set up the info with the rvec class. It makes it easy to pass across functions
   if(TYPEOF(x) == VECSXP){
     K = Rf_length(x);
     for(int k=0; k<K; ++k){
-      IndexInputVector rvec(VECTOR_ELT(x, k));
-      all_vecs.push_back(rvec);
+      std::shared_ptr<IndexInputVector> prvec = std::make_shared<IndexInputVector>(VECTOR_ELT(x, k));
+      all_pvecs.push_back(prvec);
       
       if(k == 0){
         n = Rf_length(VECTOR_ELT(x, 0));
       } else if((size_t) Rf_length(VECTOR_ELT(x, k)) != n){
-        Rf_error("All the vectors to turn into an index must be of the same length. This is currently not the case.");
+        Rcpp::stop("All the vectors to turn into an index must be of the same length. This is currently not the case.");
       }
     }
     
   } else {
-    IndexInputVector rvec(x);
-    all_vecs.push_back(rvec);
+    std::shared_ptr<IndexInputVector> prvec = std::make_shared<IndexInputVector>(x);
+    all_pvecs.push_back(prvec);
   }
   
-  return all_vecs;
+  return all_pvecs;
 }
 
 void to_index_main(const SEXP &x, IndexedVector &output, 
                    const bool do_sum, const double *p_vec_to_sum){
   
-  std::vector<IndexInputVector> all_vecs = SEXP_to_vec_r_vector(x);
+  std::vector<std::shared_ptr<IndexInputVector>> all_pvecs = SEXP_to_vec_r_vector(x);
   
-  to_index_main(all_vecs, output, do_sum, p_vec_to_sum);
+  to_index_main(all_pvecs, output, do_sum, p_vec_to_sum);
 }
 
 void to_index_main(const SEXP &x, IndexedVector &output){
@@ -956,26 +955,27 @@ void to_index_main(const SEXP &x, IndexedVector &output){
   to_index_main(x, output, do_sum, p_vec_to_sum);
 }
 
-void to_index_main(const IndexInputVector &x, IndexedVector &output, 
+void to_index_main(const std::shared_ptr<IndexInputVector> &x, IndexedVector &output, 
                    const bool do_sum, const double *p_vec_to_sum){
-  std::vector<IndexInputVector> all_vecs;
-  all_vecs.push_back(x);
-  to_index_main(all_vecs, output, do_sum, p_vec_to_sum);
+  std::vector<std::shared_ptr<IndexInputVector>> all_pvecs;
+  all_pvecs.push_back(x);
+  to_index_main(all_pvecs, output, do_sum, p_vec_to_sum);
 }
 
-void to_index_main(const IndexInputVector &x, IndexedVector &output){
+void to_index_main(const std::shared_ptr<IndexInputVector> &x, IndexedVector &output){
   const bool do_sum = false;
   const double *p_vec_to_sum = nullptr;
   to_index_main(x, output, do_sum, p_vec_to_sum);
 }
 
-void to_index_main(const std::vector<IndexInputVector> &x, IndexedVector &output){
+void to_index_main(const std::vector<std::shared_ptr<IndexInputVector>> &x, IndexedVector &output){
   const bool do_sum = false;
   const double *p_vec_to_sum = nullptr;
   to_index_main(x, output, do_sum, p_vec_to_sum);
 }
 
-void to_index_main(const std::vector<IndexInputVector> &all_vecs, IndexedVector &output, 
+void to_index_main(const std::vector<std::shared_ptr<IndexInputVector>> &all_pvecs, 
+                   IndexedVector &output, 
                    const bool do_sum, const double *p_vec_to_sum){
   
   // x: vector or list of vectors of the same length (n)
@@ -983,11 +983,11 @@ void to_index_main(const std::vector<IndexInputVector> &all_vecs, IndexedVector 
   // - index: vector of length n, from 1 to the numberof unique values of x (g)
   // - first_obs: vector of length g of the first observation belonging to each group
   
-  int K = all_vecs.size();
-  int n = all_vecs.at(0).size();
+  int K = all_pvecs.size();
+  int n = all_pvecs.at(0)->size();
   
   if(n != output.size()){
-    Rf_error("Internal error `to_index_main`: The index size allocated in output is different from the input!");
+    Rcpp::stop("Internal error `to_index_main`: The index size allocated in output is different from the input!");
   }
   
   
@@ -1005,7 +1005,7 @@ void to_index_main(const std::vector<IndexInputVector> &all_vecs, IndexedVector 
   int sum_bin_ranges = 0;
   vector<int> id_fast_int;
   for(int k=0 ; k<K ; ++k){
-    const IndexInputVector *x = &all_vecs[k];
+    const IndexInputVector *x = all_pvecs[k].get();
     if(x->is_fast_int){
       int new_bin_range = sum_bin_ranges + x->x_range_bin;
       if(new_bin_range < 17 || (K >= 2 && new_bin_range <= power_of_two(5 * n))){
@@ -1029,7 +1029,7 @@ void to_index_main(const std::vector<IndexInputVector> &all_vecs, IndexedVector 
     init_done = true;
     
     is_final = (size_t) K == id_fast_int.size();
-    multiple_ints_to_index(all_vecs, id_fast_int, p_index, n_groups, is_final, 
+    multiple_ints_to_index(all_pvecs, id_fast_int, p_index, n_groups, is_final, 
                            vec_firstobs, vec_table, vec_sum, do_sum, p_vec_to_sum);
   }
   
@@ -1057,7 +1057,7 @@ void to_index_main(const std::vector<IndexInputVector> &all_vecs, IndexedVector 
       all_k_left.erase(all_k_left.begin());
       
       is_final = all_k_left.empty();
-      general_type_to_index_single(&all_vecs[k0], p_index, n_groups, is_final, 
+      general_type_to_index_single(all_pvecs[k0].get(), p_index, n_groups, is_final, 
                                    vec_firstobs, vec_table, vec_sum, do_sum, p_vec_to_sum);
     }
     
@@ -1072,12 +1072,12 @@ void to_index_main(const std::vector<IndexInputVector> &all_vecs, IndexedVector 
         int k = all_k_left[ind];
         is_final = ind == all_k_left.size() - 1;
         if(is_res_updated_index){
-          general_type_to_index_double(&all_vecs[k], p_index, p_extra_index, n_groups, 
+          general_type_to_index_double(all_pvecs[k].get(), p_index, p_extra_index, n_groups, 
                                        is_final, vec_firstobs, vec_table, vec_sum, 
                                        do_sum, p_vec_to_sum);
           is_res_updated_index = false;
         } else {
-          general_type_to_index_double(&all_vecs[k], p_extra_index, p_index, n_groups, 
+          general_type_to_index_double(all_pvecs[k].get(), p_extra_index, p_index, n_groups, 
                                        is_final, vec_firstobs, vec_table, vec_sum, 
                                        do_sum, p_vec_to_sum);
           is_res_updated_index = true;
@@ -1092,16 +1092,16 @@ void to_index_main(const std::vector<IndexInputVector> &all_vecs, IndexedVector 
   }
 }
  
-}
+} // end namespace indexthis
  
  
  // [[Rcpp::export]]
 SEXP cpp_to_index(SEXP &x){
   
   // x can be a vector or a list of vectors of the same length
-  std::vector<indexthis::IndexInputVector> all_vecs = indexthis::SEXP_to_vec_r_vector(x);
+  std::vector<std::shared_ptr<indexthis::IndexInputVector>> all_pvecs = indexthis::SEXP_to_vec_r_vector(x);
   
-  const int n = all_vecs.at(0).size();
+  const int n = all_pvecs.at(0)->size();
   SEXP index = PROTECT(Rf_allocVector(INTSXP, n));
   
   indexthis::IndexedVector index_info(index);
@@ -1110,7 +1110,7 @@ SEXP cpp_to_index(SEXP &x){
   // computing the index 
   //
   
-  to_index_main(all_vecs, index_info);
+  to_index_main(all_pvecs, index_info);
   
   //
   // saving the results 
@@ -1130,13 +1130,6 @@ SEXP cpp_to_index(SEXP &x){
   Rf_setAttrib(res, R_NamesSymbol, indexthis::std_string_to_r_string({"index", "first_obs", "table"}));
     
   UNPROTECT(2);
-  
-  // we unprotect if we have converted some vectors to character
-  for(const auto &v : all_vecs){
-    if(v.is_protect){
-      UNPROTECT(2);
-    }
-  }
   
   return res;